requirements.txt
# Essential packages for PGU-net project
numpy>=1.26.4,<2.0.0
matplotlib==3.10.5
scikit-image==0.24.0
scipy==1.13.1
pyparsing==3.1.4

# pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu128

# Additional packages for visualization and utilities
Pillow>=8.0.0
tqdm>=4.60.0
tensorboard==2.17.0
opencv-python>=4.5.0
pathlib
scikit-learn
seaborn
wandb
albumentations